# api/openapi.yaml
openapi: 3.1.0
info:
  title: github-export local API
  version: "1.0.0"
  description: >
    A GitHub-compatible read mirror served from a local SQLite store and a local
    git clone. Mirrored reads are answered offline; any unmirrored read and all
    writes fall through to api.github.com via the built-in proxy. Paths and JSON
    field names match GitHub's REST API so existing clients (including `gh`) work
    unchanged. The server binds localhost and requires no token from clients.
servers:
  - url: http://localhost:8080
    description: Local serve address (default)
...
paths:
...
  /search/issues:
    get:
      tags: [search]
      summary: Search issues and pull requests
      parameters:
        - { name: q, in: query, required: true, schema: { type: string }, description: "GitHub search query (is:, label:, author:, state:, milestone:, free text)" }
        - $ref: '#/components/parameters/perPage'
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_count: { type: integer }
                  incomplete_results: { type: boolean }
                  items: { type: array, items: { type: object } }
...
components:
  parameters:
...
    perPage: { name: per_page, in: query, schema: { type: integer, default: 30, maximum: 100 } }
    page: { name: page, in: query, schema: { type: integer, default: 1 } }
...